moved BaseEntity
