+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/entities/Agent.kt
@@ -1,5 +1,6 @@
 package com.saveourtool.save.entities
 
+import com.saveourtool.save.spring.entity.BaseEntity
 import javax.persistence.Entity
 import javax.persistence.FetchType
 import javax.persistence.JoinColumn
