+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/entities/OriginalLogin.kt
@@ -1,6 +1,9 @@
 package com.saveourtool.save.entities
 
+import com.saveourtool.save.spring.entity.BaseEntity
+
 import com.fasterxml.jackson.annotation.JsonBackReference
+
 import javax.persistence.Entity
 import javax.persistence.JoinColumn
 import javax.persistence.ManyToOne
