+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/entities/Test.kt
@@ -1,5 +1,6 @@
 package com.saveourtool.save.entities
 
+import com.saveourtool.save.spring.entity.BaseEntity
 import com.saveourtool.save.test.TestDto
 import com.saveourtool.save.utils.DATABASE_DELIMITER
 import java.time.LocalDateTime
