+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/entities/TestExecution.kt
@@ -2,6 +2,7 @@ package com.saveourtool.save.entities
 
 import com.saveourtool.save.agent.TestExecutionDto
 import com.saveourtool.save.domain.TestResultStatus
+import com.saveourtool.save.spring.entity.BaseEntity
 
 import java.time.LocalDateTime
 import java.time.ZoneOffset
