+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/entities/benchmarks/AwesomeBenchmarks.kt
@@ -1,6 +1,6 @@
 package com.saveourtool.save.entities.benchmarks
 
-import com.saveourtool.save.entities.BaseEntity
+import com.saveourtool.save.spring.entity.BaseEntity
 import com.saveourtool.save.utils.DATABASE_DELIMITER
 import javax.persistence.Entity
 import javax.persistence.EnumType
