+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/spring/entity/BaseEntity.kt
@@ -1,4 +1,4 @@
-package com.saveourtool.save.entities
+package com.saveourtool.save.spring.entity
 
 import com.saveourtool.save.validation.Validatable
 import javax.persistence.GeneratedValue
