+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/spring/entity/BaseEntityWithDto.kt
@@ -1,4 +1,6 @@
-package com.saveourtool.save.entities
+package com.saveourtool.save.spring.entity
+
+import com.saveourtool.save.entities.DtoWithId
 
 /**
  * base class for all entities with DTO with [D] type
