+++ save-cloud-common/src/jvmMain/kotlin/com/saveourtool/save/spring/repository/BaseEntityRepository.kt
@@ -1,6 +1,6 @@
 package com.saveourtool.save.spring.repository
 
-import com.saveourtool.save.entities.BaseEntity
+import com.saveourtool.save.spring.entity.BaseEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 import org.springframework.data.repository.NoRepositoryBean
